Enum liquid::Value

pub enum Value {
    Num(f32),
    Str(String),
    Object(HashMap<String, Value>),
    Array(Vec<Value>),
    Bool(bool),
}

An enum to represent different value types

Variants
Num(f32)Str(String)Object(HashMap<String, Value>)Array(Vec<Value>)Bool(bool)
Methods

impl<'a> Value

fn str(val: &str) -> Value

Shorthand function to create Value::Str from a string slice.

fn as_str(&'a self) -> Option<&'a str>

fn from_token(t: &Token) -> Result<Value, Error>

Parses a token that can possibly represent a Value to said Value. Returns an Err if the token can not be interpreted as a Value.