liquid

Module model

Liquid data model.

Modules

map

Type representing a Liquid object, payload of the Value::Object variant

Macros

array

A value::Array literal.

object

A value::Object literal.

scalar

A value::Scalar literal.

value

A value::Value literal.

Structs

Date

Liquid’s native date only type.

DateTime

Liquid’s native date + time type.

KStringRef

A reference to a UTF-8 encoded, immutable string.

Object

Type representing a Liquid object, payload of the Value::Object variant

ObjectRender

Helper for ObjectView::render

ObjectSource

Helper for ObjectView::source

Path

Path to a value in an Object.

PathIter

Iterate over indexes in a Value’s Path.

ScalarCow

A Liquid scalar value

ValueViewCmp

Value comparison semantics for types implementing ValueView.

Enums

DisplayCow

Abstract the lifetime of a Display.

State

Queryable state for a Value.

Value

An enum to represent different value types

ValueCow

Abstract the lifetime of a Value.

Traits

ArrayView

Accessor for arrays.

ObjectIndex

Owned object index

ObjectView

Accessor for objects.

ValueView

Accessor for Values.

_ObjectView

Accessor for objects.

_ValueView

Accessor for Values.

Functions

find

Find a ValueView nested in an ObjectView

from_value

Convert a value into T.

to_object

Convert a T into liquid_core::model::Object.

to_scalar

Convert a T into liquid_core::model::Scalar.

to_value

Convert a T into liquid_core::model::Value.

try_find

Find a ValueView nested in an ObjectView

Type Aliases

Array

Type representing a Liquid array, payload of the Value::Array variant

KString

A UTF-8 encoded, immutable string.

KStringCow

A reference to a UTF-8 encoded, immutable string.

Scalar

A Liquid scalar value

Derive Macros

ObjectView

ValueView