liquid::model

Trait _ValueView

pub trait _ValueView: Debug {

    // Required methods
    fn as_debug(&self) -> &dyn Debug;
    fn render(&self) -> DisplayCow<'_>;
    fn source(&self) -> DisplayCow<'_>;
    fn type_name(&self) -> &'static str;
    fn query_state(&self, state: State) -> bool;
    fn to_kstr(&self) -> KStringCowBase<'_>;
    fn to_value(&self) -> Value;

    // Provided methods
    fn as_scalar(&self) -> Option<ScalarCow<'_>> { ... }
    fn is_scalar(&self) -> bool { ... }
    fn as_array(&self) -> Option<&dyn ArrayView> { ... }
    fn is_array(&self) -> bool { ... }
    fn as_object(&self) -> Option<&dyn ObjectView> { ... }
    fn is_object(&self) -> bool { ... }
    fn as_state(&self) -> Option<State> { ... }
    fn is_state(&self) -> bool { ... }
    fn is_nil(&self) -> bool { ... }

}

Accessor for Values.

Required Methods

fn as_debug(&self) -> &dyn Debug

Get a Debug representation

fn render(&self) -> DisplayCow<'_>

A Display for a BoxedValue rendered for the user.

fn source(&self) -> DisplayCow<'_>

A Display for a Value as source code.

fn type_name(&self) -> &'static str

Report the data type (generally for error reporting).

fn query_state(&self, state: State) -> bool

Query the value’s state

fn to_kstr(&self) -> KStringCowBase<'_>

Interpret as a string.

fn to_value(&self) -> Value

Convert to an owned type.

Provided Methods

fn as_scalar(&self) -> Option<ScalarCow<'_>>

Extracts the scalar value if it is a scalar.

fn is_scalar(&self) -> bool

Tests whether this value is a scalar

fn as_array(&self) -> Option<&dyn ArrayView>

Extracts the array value if it is an array.

fn is_array(&self) -> bool

Tests whether this value is an array

fn as_object(&self) -> Option<&dyn ObjectView>

Extracts the object value if it is a object.

fn is_object(&self) -> bool

Tests whether this value is an object

fn as_state(&self) -> Option<State>

Extracts the state if it is one

fn is_state(&self) -> bool

Tests whether this value is state

fn is_nil(&self) -> bool

Tests whether this value is nil

See the Stack overflow table