Function liquid::parser::consume_value_token

pub fn consume_value_token(tokens: &mut Iter<Token>) -> Result<Token, Error>

Extracts a token from the token stream that can be used to express a value. For our purposes, this is either a string literal, number literal or an identifier that might refer to a variable.