Struct liquid::LiquidOptions

pub struct LiquidOptions {
    pub blocks: HashMap<String, Box<Block>>,
    pub tags: HashMap<String, Box<Tag>>,
    pub file_system: Option<PathBuf>,
}

Options that liquid::parse takes

Fields
blocks: HashMap<String, Box<Block>>

Holds all custom block-size tags

tags: HashMap<String, Box<Tag>>

Holds all custom tags

file_system: Option<PathBuf>

The path to which paths in include tags should be relative to

Methods

impl LiquidOptions

fn with_known_blocks() -> LiquidOptions

Creates a LiquidOptions instance, pre-seeded with all known tags and blocks.

fn register_known_blocks(&mut self)

Registers all known tags and blocks in an existing options struct

fn register_block(&mut self, name: &str, block: Box<Block>)

Inserts a new custom block into the options object

fn register_tag(&mut self, name: &str, tag: Box<Tag>)

Inserts a new custom tag into the options object